login

+++ cake2/app/Controller/UsersController.php
@@ -1,39 +1,24 @@
 <?php
 
-App::uses('AppController', 'Controller');
-
 class UsersController extends AppController {
-	public $scaffold = 'painel';
-
-    protected function _isPrefix($prefix) {
-        return isset($this->params['prefix']) &&
-        $this->params['prefix'] === $prefix;
-        }
 
     public function beforeFilter() {
         parent::beforeFilter();
-        if ($this->_isPrefix('painel'))
-        $this->layout = 'painel'; // Layout padrão do prefixo
-		// Allow users to register and logout.
-        $this->Auth->allow('add', 'logout');
-        if (!$this->_isPrefix('painel'))
-        $this->Auth->allow();
-	}
-
-
-
-	public function painel_login() {
-		if ($this->request->is('post')) {
-			if ($this->Auth->login()) {
-				return $this->redirect($this->Auth->redirectUrl());
-			}
-			$this->Flash->error(__('Invalid username or password, try again'));
-		}
-	}
-
-	public function logout() {
-		return $this->redirect($this->Auth->logout());
-	}
+        $this->Auth->allow('add'); // Permitindo que os usuários se registrem
+    }
+    
+    public function login() {
+        if ($this->Auth->login()) {
+            $this->redirect($this->Auth->redirect());
+        } else {
+            $this->Flash->error(__('Invalid username or password, try again'));
+        }
+    }
+    
+    public function logout() {
+        $this->redirect($this->Auth->logout());
+    }
+    
 
     public function index() {
         $this->User->recursive = 0;
@@ -41,8 +26,7 @@ class UsersController extends AppController {
     }
 
     public function view($id = null) {
-        $this->User->id = $id;
-        if (!$this->User->exists()) {
+        if (!$this->User->exists($id)) {
             throw new NotFoundException(__('Invalid user'));
         }
         $this->set('user', $this->User->findById($id));
@@ -53,11 +37,10 @@ class UsersController extends AppController {
             $this->User->create();
             if ($this->User->save($this->request->data)) {
                 $this->Flash->success(__('The user has been saved'));
-                return $this->redirect(array('action' => 'index'));
+                $this->redirect(array('action' => 'index'));
+            } else {
+                $this->Flash->error(__('The user could not be saved. Please, try again.'));
             }
-            $this->Flash->error(
-                __('The user could not be saved. Please, try again.')
-            );
         }
     }
 
@@ -69,11 +52,10 @@ class UsersController extends AppController {
         if ($this->request->is('post') || $this->request->is('put')) {
             if ($this->User->save($this->request->data)) {
                 $this->Flash->success(__('The user has been saved'));
-                return $this->redirect(array('action' => 'index'));
+                $this->redirect(array('action' => 'index'));
+            } else {
+                $this->Flash->error(__('The user could not be saved. Please, try again.'));
             }
-            $this->Flash->error(
-                __('The user could not be saved. Please, try again.')
-            );
         } else {
             $this->request->data = $this->User->findById($id);
             unset($this->request->data['User']['password']);
@@ -81,21 +63,20 @@ class UsersController extends AppController {
     }
 
     public function delete($id = null) {
-        // Prior to 2.5 use
-        // $this->request->onlyAllow('post');
-
-        $this->request->allowMethod('post');
-
+        if (!$this->request->is('post')) {
+            throw new MethodNotAllowedException();
+        }
         $this->User->id = $id;
         if (!$this->User->exists()) {
             throw new NotFoundException(__('Invalid user'));
         }
         if ($this->User->delete()) {
             $this->Flash->success(__('User deleted'));
-            return $this->redirect(array('action' => 'index'));
+            $this->redirect(array('action' => 'index'));
         }
         $this->Flash->error(__('User was not deleted'));
-        return $this->redirect(array('action' => 'index'));
+        $this->redirect(array('action' => 'index'));
     }
+
 
 }
